+++ R/rules-quickr.R
@@ -1205,7 +1205,7 @@ quickr_emit_reduce <- function(kind, out_sym, operand_expr, shape_in, axes, drop
       1L
     }
     if (as.character(dtype(out_aval)) %in% "bool") {
-      cli_abort("{kind}: pred reductions are not supported by quickr lowering")
+      cli_abort("{kind}: reductions of a boolean are not supported by quickr lowering")
     }
   } else {
     init_acc_scalar <- NULL
@@ -2017,7 +2017,7 @@ local({
 
       if (dt_lhs %in% "bool" || dt_rhs %in% "bool") {
         if (!prim_name %in% c("equal", "not_equal")) {
-          cli_abort("{prim_name}: comparisons on {.val pred} values are not supported by quickr lowering")
+          cli_abort("{prim_name}: comparisons on {.val bool} values are not supported by quickr lowering")
         }
 
         a <- inputs[[1L]]
@@ -2059,7 +2059,7 @@ local({
     function(prim_name, inputs, params, out_syms, input_nodes, out_avals, ctx = NULL) {
       dt <- as.character(dtype(input_nodes[[1L]]$aval))
       if (!dt %in% "bool") {
-        cli_abort("{prim_name}: only {.val pred} dtype is supported by quickr lowering")
+        cli_abort("{prim_name}: only the {.val bool} data type is supported by quickr lowering")
       }
 
       a <- inputs[[1L]]
@@ -2087,7 +2087,7 @@ local({
     function(prim_name, inputs, params, out_syms, input_nodes, out_avals, ctx = NULL) {
       dt <- as.character(dtype(input_nodes[[1L]]$aval))
       if (!dt %in% "bool") {
-        cli_abort("not: only {.val pred} dtype is supported by quickr lowering")
+        cli_abort("not: only the {.val bool} data type is supported by quickr lowering")
       }
       quickr_emit_assign(out_syms[[1L]], rlang::call2("!", inputs[[1L]]))
     }
@@ -2281,7 +2281,7 @@ local({
       operand_node <- input_nodes[[1L]]
       dt_in <- as.character(dtype(operand_node$aval))
       if (!dt_in %in% "bool") {
-        cli_abort("reduce_any: only {.val pred} inputs are supported by quickr lowering")
+        cli_abort("reduce_any: only a {.val bool} input is supported by quickr lowering")
       }
       quickr_emit_reduce_boolean(
         "any",
@@ -2303,7 +2303,7 @@ local({
       operand_node <- input_nodes[[1L]]
       dt_in <- as.character(dtype(operand_node$aval))
       if (!dt_in %in% "bool") {
-        cli_abort("reduce_all: only {.val pred} inputs are supported by quickr lowering")
+        cli_abort("reduce_all: only a {.val bool} input is supported by quickr lowering")
       }
       quickr_emit_reduce_boolean(
         "all",